Feat(#22): Todo 컴포넌트 정리 및 스토리북 작성
